bpf_loader: extract vm execute to program-runtime
