+++ program-runtime/src/lib.rs
@@ -22,12 +22,16 @@ pub mod memory;
 pub mod serialization;
 pub mod stable_log;
 pub mod sysvar_cache;
+pub mod vm;
 
 // re-exports for macros
 pub mod __private {
     pub use {
-        solana_account::ReadableAccount, solana_hash::Hash,
-        solana_instruction::error::InstructionError, solana_rent::Rent,
+        crate::vm::{MEMORY_POOL, calculate_heap_cost, create_vm},
+        solana_account::ReadableAccount,
+        solana_hash::Hash,
+        solana_instruction::error::InstructionError,
+        solana_rent::Rent,
         solana_transaction_context::TransactionContext,
     };
 }
