+++ programs/bpf_loader/Cargo.toml
@@ -20,7 +20,7 @@ name = "solana_bpf_loader_program"
 default = ["metrics"]
 agave-unstable-api = []
 metrics = ["solana-program-runtime/metrics"]
-sbpf-debugger = ["solana-sbpf/debugger"]
+sbpf-debugger = ["solana-program-runtime/sbpf-debugger", "solana-sbpf/debugger"]
 shuttle-test = [
     "solana-program-runtime/shuttle-test",
     "solana-sbpf/shuttle-test",
@@ -31,7 +31,6 @@ svm-internal = []
 [dependencies]
 agave-syscalls = { workspace = true }
 bincode = { workspace = true }
-cfg-if = { workspace = true }
 qualifier_attr = { workspace = true }
 solana-account = { workspace = true }
 solana-bincode = { workspace = true }
